requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-decouple==3.8
psycopg2-binary==2.9.6
gunicorn==20.1.0
Pillow==9.5.0
google-generativeai==0.3.0
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
requests==2.31.0
dj-database-url==1.3.0
whitenoise==6.5.0
